move constant to utils
